Refactor MTF coordination services
